linemon_frontend/requirements.txt
fastapi>=0.116.0
uvicorn>=0.35.0
jinja2>=3.1.4
httpx>=0.28.1
PyMySQL>=1.1.1
python-dotenv>=1.1.1

